chore: Update releases UI actions
