--- static/js/publisher/pages/Releases/components/releasesConfirmDetails/globalRow.js
@@ -1,29 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-const GlobalRow = ({ useGlobal, toggleGlobal }) => {
-  return (
-    <div className="p-release-details-row is-global">
-      <span className="p-release-details-row__global-check">
-        <input
-          type="checkbox"
-          checked={useGlobal}
-          onChange={toggleGlobal}
-          id="useGlobalToggle"
-        />
-        <label htmlFor="useGlobalToggle">
-          Use this percentage for all new progressive releases
-        </label>
-      </span>
-    </div>
-  );
-};
-
-GlobalRow.propTypes = {
-  useGlobal: PropTypes.bool,
-  toggleGlobal: PropTypes.func,
-  globalPercentage: PropTypes.number,
-  updatePercentage: PropTypes.func,
-};
-
-export default GlobalRow;
