+++ static/sass/_snapcraft_p-release-details-row.scss
@@ -1,47 +1,20 @@
 @mixin snapcraft-p-release-details-row {
   .p-release-details-row {
     align-items: center;
-    display: grid;
-    grid-column-gap: 1rem;
-    grid-template-columns: 4rem 15rem 2rem auto 16rem;
+    display: flex;
+    gap: 0.25rem;
     padding: 0.25rem 0;
-
-    @media screen and (max-width: $breakpoint-large) {
-      grid-column-gap: 0.5rem;
-      grid-template-columns: 3rem 11rem 2rem auto 16rem;
-    }
 
     // Same as the global nav breakpoint
     @media screen and (width <= 875px) {
-      grid-template-columns: 3rem 11rem 2rem auto 0;
-
       &__notes {
         display: none;
       }
     }
 
-    &.is-closing {
-      grid-template-columns: 4rem auto;
-
-      @media screen and (max-width: $breakpoint-large) {
-        grid-template-columns: 3rem auto;
-      }
-    }
-
-    &.is-global {
-      padding: 0.25rem;
-
-      .p-release-details-row__global-check {
-        grid-column-start: 4;
-      }
-    }
-
-    &__join {
-      text-align: center;
-    }
-
     &__progress {
       display: grid;
+      flex: 1;
       font-weight: 400;
       grid-column-gap: 0.5rem;
       grid-template-columns: auto 8rem;
@@ -56,5 +29,14 @@
     border-top: 1px dashed $color-mid-light;
     margin-top: 0.5rem;
     padding-top: 0.5rem;
+  }
+
+  .progressive-release-control {
+    padding-top: 1rem;
+  }
+
+  .progressive-release-control__inner {
+    display: flex;
+    gap: 1rem;
   }
 }
